+++ CHANGELOG.md
@@ -34,6 +34,10 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 
 #### :nail_care: Polish
 
+* Improved documentation:
+  * `bRouter`
+  * `super/i-block/modules/state`
+
 * Added documentation:
   * `bCheckbox`
   * `bRadioButton`